// src/app/api/ui/models/nic-registry-grid.interface.ts
export interface NicRegistryGrid {
  contractDate: string;
  contractNumber: string;
  createDate: string;
  customerInn: string;
  customerKpp: string;
  customerName: string;
  id: number;
  isAvailableForCurrentOrganization: boolean;
  participantEmail: string;
  participantInn: string;
  participantKpp: string;
  participantName: string;
  reasonForAdding: string;
  responsibleEmail: string;
  responsibleFullName: string;
  responsiblePhoneNumber: string;
  tenantName: string;
  tradeId: number;
  tradeName: string;
}