// src/app/api/swagger/models/api-nic-registry-grid-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface ApiNicRegistryGridDto {
  ContractDate?: string | null;
  ContractNumber?: string | null;
  CreateDate?: string;
  CustomerInn?: string | null;
  CustomerKpp?: string | null;
  CustomerName?: string | null;
  Id?: number;
  IsAvailableForCurrentOrganization?: boolean;
  ParticipantEmail?: string | null;
  ParticipantInn?: string | null;
  ParticipantKpp?: string | null;
  ParticipantName?: string | null;
  ReasonForAdding?: string | null;
  ResponsibleEmail?: string | null;
  ResponsibleFullName?: string | null;
  ResponsiblePhoneNumber?: string | null;
  TenantName?: string | null;
  TradeId?: number | null;
  TradeName?: string | null;
}